+++ bitswap/server/server.go
@@ -20,8 +20,6 @@ import (
 	"github.com/ipfs/go-cid"
 	logging "github.com/ipfs/go-log/v2"
 	"github.com/ipfs/go-metrics-interface"
-	process "github.com/jbenet/goprocess"
-	procctx "github.com/jbenet/goprocess/context"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"go.uber.org/zap"
 )
@@ -57,7 +55,12 @@ type Server struct {
 	// the total number of simultaneous threads sending outgoing messages
 	taskWorkerCount int
 
-	process process.Process
+	// Cancel stops the server
+	cancel    context.CancelFunc
+	closing   chan struct{}
+	closeOnce sync.Once
+	// waitWorkers waits for all worker goroutines to exit.
+	waitWorkers sync.WaitGroup
 
 	// newBlocks is a channel for newly added blocks to be provided to the
 	// network.  blocks pushed down this channel get buffered and fed to the
@@ -78,20 +81,13 @@ type Server struct {
 func New(ctx context.Context, network bsnet.BitSwapNetwork, bstore blockstore.Blockstore, options ...Option) *Server {
 	ctx, cancel := context.WithCancel(ctx)
 
-	px := process.WithTeardown(func() error {
-		return nil
-	})
-	go func() {
-		<-px.Closing() // process closes first
-		cancel()
-	}()
-
 	s := &Server{
 		sentHistogram:      bmetrics.SentHist(ctx),
 		sendTimeHistogram:  bmetrics.SendTimeHist(ctx),
 		taskWorkerCount:    defaults.BitswapTaskWorkerCount,
 		network:            network,
-		process:            px,
+		cancel:             cancel,
+		closing:            make(chan struct{}),
 		provideEnabled:     true,
 		hasBlockBufferSize: defaults.HasBlockBufferSize,
 		provideKeys:        make(chan cid.Cid, provideKeysBufferSize),
@@ -103,7 +99,6 @@ func New(ctx context.Context, network bsnet.BitSwapNetwork, bstore blockstore.Bl
 	}
 
 	s.engine = decision.NewEngine(
-		ctx,
 		bstore,
 		network.ConnectionManager(),
 		network.Self(),
@@ -111,7 +106,7 @@ func New(ctx context.Context, network bsnet.BitSwapNetwork, bstore blockstore.Bl
 	)
 	s.engineOptions = nil
 
-	s.startWorkers(ctx, px)
+	s.startWorkers(ctx)
 
 	return s
 }
@@ -293,33 +288,31 @@ func (bs *Server) WantlistForPeer(p peer.ID) []cid.Cid {
 	return out
 }
 
-func (bs *Server) startWorkers(ctx context.Context, px process.Process) {
-	bs.engine.StartWorkers(ctx, px)
-
+func (bs *Server) startWorkers(ctx context.Context) {
 	// Start up workers to handle requests from other nodes for the data on this node
+	bs.waitWorkers.Add(bs.taskWorkerCount)
 	for i := 0; i < bs.taskWorkerCount; i++ {
 		i := i
-		px.Go(func(px process.Process) {
-			bs.taskWorker(ctx, i)
-		})
+		go bs.taskWorker(ctx, i)
 	}
 
 	if bs.provideEnabled {
-		// Start up a worker to manage sending out provides messages
-		px.Go(func(px process.Process) {
-			bs.provideCollector(ctx)
-		})
+		bs.waitWorkers.Add(1)
+		go bs.provideCollector(ctx)
 
 		// Spawn up multiple workers to handle incoming blocks
 		// consider increasing number if providing blocks bottlenecks
 		// file transfers
-		px.Go(bs.provideWorker)
+		bs.waitWorkers.Add(1)
+		go bs.provideWorker(ctx)
 	}
 }
 
 func (bs *Server) taskWorker(ctx context.Context, id int) {
-	defer log.Debug("bitswap task worker shutting down...")
+	defer bs.waitWorkers.Done()
+
 	log := log.With("ID", id)
+	defer log.Debug("bitswap task worker shutting down...")
 	for {
 		log.Debug("Bitswap.TaskWorker.Loop")
 		select {
@@ -341,8 +334,7 @@ func (bs *Server) taskWorker(ctx context.Context, id int) {
 				}
 				bs.sendBlocks(ctx, envelope)
 
-				dur := time.Since(start)
-				bs.sendTimeHistogram.Observe(dur.Seconds())
+				bs.sendTimeHistogram.Observe(time.Since(start).Seconds())
 
 			case <-ctx.Done():
 				return
@@ -452,7 +444,7 @@ func (bs *Server) Stat() (Stat, error) {
 // that those blocks are available in the blockstore before calling this function.
 func (bs *Server) NotifyNewBlocks(ctx context.Context, blks ...blocks.Block) error {
 	select {
-	case <-bs.process.Closing():
+	case <-bs.closing:
 		return errors.New("bitswap is closed")
 	default:
 	}
@@ -466,8 +458,8 @@ func (bs *Server) NotifyNewBlocks(ctx context.Context, blks ...blocks.Block) err
 			select {
 			case bs.newBlocks <- blk.Cid():
 				// send block off to be reprovided
-			case <-bs.process.Closing():
-				return bs.process.Close()
+			case <-bs.closing:
+				return nil
 			}
 		}
 	}
@@ -476,6 +468,7 @@ func (bs *Server) NotifyNewBlocks(ctx context.Context, blks ...blocks.Block) err
 }
 
 func (bs *Server) provideCollector(ctx context.Context) {
+	defer bs.waitWorkers.Done()
 	defer close(bs.provideKeys)
 	var toProvide []cid.Cid
 	var nextKey cid.Cid
@@ -508,18 +501,16 @@ func (bs *Server) provideCollector(ctx context.Context) {
 	}
 }
 
-func (bs *Server) provideWorker(px process.Process) {
-	// FIXME: OnClosingContext returns a _custom_ context type.
-	// Unfortunately, deriving a new cancelable context from this custom
-	// type fires off a goroutine. To work around this, we create a single
-	// cancelable context up-front and derive all sub-contexts from that.
-	//
-	// See: https://github.com/ipfs/go-ipfs/issues/5810
-	ctx := procctx.OnClosingContext(px)
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
+func (bs *Server) provideWorker(ctx context.Context) {
 	limit := make(chan struct{}, provideWorkerMax)
+	defer func() {
+		// Wait until all limitGoProvide goroutines are done before declaring
+		// this worker as done.
+		for i := 0; i < provideWorkerMax; i++ {
+			limit <- struct{}{}
+		}
+		bs.waitWorkers.Done()
+	}()
 
 	limitedGoProvide := func(k cid.Cid, wid int) {
 		defer func() {
@@ -540,25 +531,18 @@ func (bs *Server) provideWorker(px process.Process) {
 
 	// worker spawner, reads from bs.provideKeys until it closes, spawning a
 	// _ratelimited_ number of workers to handle each key.
-	for wid := 2; ; wid++ {
+	wid := 2
+	for k := range bs.provideKeys {
 		log.Debug("Bitswap.ProvideWorker.Loop")
-
 		select {
-		case <-px.Closing():
+		case limit <- struct{}{}:
+			go limitedGoProvide(k, wid)
+		case <-ctx.Done():
 			return
-		case k, ok := <-bs.provideKeys:
-			if !ok {
-				log.Debug("provideKeys channel closed")
-				return
-			}
-			select {
-			case <-px.Closing():
-				return
-			case limit <- struct{}{}:
-				go limitedGoProvide(k, wid)
-			}
 		}
+		wid++
 	}
+	log.Debug("provideKeys channel closed")
 }
 
 func (bs *Server) ReceiveMessage(ctx context.Context, p peer.ID, incoming message.BitSwapMessage) {
@@ -597,7 +581,13 @@ func (bs *Server) PeerDisconnected(p peer.ID) {
 	bs.engine.PeerDisconnected(p)
 }
 
-// Close is called to shutdown the Client
-func (bs *Server) Close() error {
-	return bs.process.Close()
+// Close is called to shutdown the Server. Returns when all workers and
+// decision engine have finished. Safe to calling multiple times/concurrently.
+func (bs *Server) Close() {
+	bs.closeOnce.Do(func() {
+		close(bs.closing)
+		bs.cancel()
+	})
+	bs.engine.Close()
+	bs.waitWorkers.Wait()
 }